fixing bugs, finishing the project

+++ AppLib/Display.h
@@ -14,7 +14,7 @@ public:
 
 	string drawConnection(int x1, int y1, int x2, int y2);
 	string drawStation(int x, int y, int index);
-	string drawTrain(int x, int y, int index);
+	string drawTrain(double x, double y, int index);
 	string addScripts();
 	string addStationInfo(int index, string name, int peopleNumber);
 	string addTrainInfo(int index, string name, int maxCapacity, int peopleNumber);
